refactor(core): exceptions
